Rewrite traffic_info using Python and ChromeDriver
The webkit2png seems to be abandoned.
Modern web browsers are now implementing headless mode and remote
browser piloting through a webdriver.

I decided to rebuild the project from scratch using Python 3 and
Selenium.

This first iteration focus on ChromeDriver but I plan to support more
webdrivers in the future.